Moving files, cleaning up things
(cherry picked from commit fae9005c93524ff57c414a4b70bd5a6178e720b4)

Conflicts:
	hpx/lcos/async_callback.hpp
	hpx/runtime/components/stubs/runtime_support.hpp
