+++ hpx/lcos/detail/async_colocated_callback_fwd.hpp
@@ -6,7 +6,7 @@
 #if !defined(HPX_LCOS_ASYNC_COLOCATED_CALLBACK_FWD_MAR_30_2015_1145PM)
 #define HPX_LCOS_ASYNC_COLOCATED_CALLBACK_FWD_MAR_30_2015_1145PM
 
-#include <hpx/lcos/async_colocated_fwd.hpp>
+#include <hpx/lcos/detail/async_colocated_fwd.hpp>
 
 namespace hpx { namespace detail
 {
