+++ src/components/component_storage/component_module.cpp
@@ -3,6 +3,8 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#include <hpx/include/apply.hpp>
+#include <hpx/include/async.hpp>
 #include <hpx/include/components.hpp>
 #include <hpx/components/component_storage/server/migrate_to_storage.hpp>
